requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2.2
django-adaptors==0.2.5
django-cors-headers==3.3.0
django-environ==0.10.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
install==1.3.5
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
promise==2.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
Rx==1.6.3
selenium==4.9.1
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
